some simplifications and merging

+++ IPRgeneration/mergeIPRGraphs.sh
@@ -5,27 +5,21 @@
 # Requires an Eventdisplay installation
 # including all environmental variables
 #
-# Change the two variables at the top:
-#  SCRATCH: scratch space to save the output
-#  MOONOPT: option to produce pedestals for halfmoon NSB or dark time
 
-SCRATCH="$CTA_USER_DATA_DIR/tmp/"
+if [ $# -lt 2 ]; then
+    echo "
+    ./produceIPRGraphs.sh <directory with data products (pedestal files)> <IPR file>
+    "
+    exit
+fi
+SCRATCH=${1}
+IPRFILE=${2}
 CDIR=$(pwd)
-
-ZE="20"
-# ZE="60"
-
-MOONOPT="" # Set to -DHALFMOON for half moon or leave empty or dark conditions (i.e., MOONOPT="")
-MOON=`echo "${MOONOPT#*-D}" | tr "[:upper:]" "[:lower:]"`
-
-[ -z "$MOON" ] && echo "Running dark conditions" || MOON="-${MOON}"
-
-IPRFILE="prod5${MOON}-ze-${ZE}-IPR.root"
 
 # list of files to be merged
 FLIST=${SCRATCH}/pedestals-${MOON}-ze-${ZE}.list
 rm -f ${FLIST}
-find $SCRATCH -name "pedestal*${MOON}-ze-${ZE}*.pedestal.root" > ${FLIST}
+find $SCRATCH -name "*.pedestal.root" > ${FLIST}
 echo "IPR files to be merged (from ${FLIST}):"
 cat ${FLIST}
 
@@ -33,7 +27,7 @@ cat ${FLIST}
 root -l -q -b 'mergeIPRGraphs.C( '\"$IPRFILE\"', '\"$FLIST\"' )'
 
 # add log files
-for logFileNow in $(ls ${SCRATCH}/pedestal*${MOON}*"ze-"${ZE}*.log)
+for logFileNow in $(ls ${SCRATCH}/*.log)
 do
     logFile=$(basename -- "$logFileNow")
     fileTitle="${logFile%.*}"
